requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
sqlalchemy
pydantic[email]
